great commit

+++ HomeProject_2/Utill.c
@@ -2,6 +2,8 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
 #include "Utill.h"
 
 char *createDynStr(const char *msg)
@@ -10,7 +12,6 @@ char *createDynStr(const char *msg)
 	char temp[255];
 	printf("Enter %s: ", msg);
 	scanf("%s", temp);
-
 	str = (char *)malloc((strlen(temp) + 1) * sizeof(char));
 	if (!str)
 		return NULL;
@@ -21,8 +22,8 @@ char *createDynStr(const char *msg)
 
 int isUpperCase(const char* str)
 {
-	int length = strlen(str);
-	for (int i = 0; i < length; i++)
+	size_t length = strlen(str);
+	for (size_t i = 0; i < length; i++)
 		if (isupper(str[i]) != 1)
 			return 0;
 	return 1;
